// src/app/ts/types/RoleStatisticsType.ts
export type RoleStatisticsType = {
    tank: {
        tournamentsPlayed: number;
        averageWinrate: number;
        lastRate: string;
    };
    dps: {
        tournamentsPlayed: number;
        averageWinrate: number;
        lastRate: string;
    };
    support: {
        tournamentsPlayed: number;
        averageWinrate: number;
        lastRate: string;
    };
};